Load trees with filter from airtable sqlite
The filter works to change the foods shown.
Not yet able to add/delete/update any trees.
This changes the food.id from a comma separated
string of integers, the old food.id,
to an array of strings, the ned food.id in airtable.
Before this change the backend could assume that
all of the tree filter values were ints passed in
a comma separated string.
Now the food.ids are a json encoded array, and the
others are still ints.