Remove vendor dir
I am not able to test mailgun from this new composer.lock
If there is a problem, we can revert this or probably better to
roll forward with a fix.
with js node_modules custom changes were made to the modules and thats
why they were being tracked in git.

For reference the installed.json in the original list the following
top level packages:
{"name":"symfony/event-dispatcher","version_normalized":"2.8.8.0"}
{"name":"guzzle/guzzle","version_normalized":"3.9.3.0"}
{"name":"mailgun/mailgun-php","version_normalized":"1.7.2.0"}