Stop calculating seasonal foods
Sqlite does not have the FIND_IN_SET used by this query.
It is also of questionable value because of its design.

This was complex query searching for all trees whose foods
where picked within two weeks of this time of year
then writing to the foods table a 0 or 1 for in season
A later function would get all foods where season = 1
to render only the trees with ready to pick food by default.

However this expensive update query runs every time the map loads
an alternative design could be to have a pick_start and pick_end
time of year that is queried and only updated when a tree for that food
is picked.