Link Trees to Food to get their icons
Trees originally linked to the food.old_id so when
the Food.id became an airtable_id the tree.food.icon
link was broken.

Once the icons are found, the field was empty in airtable.
The field currently expects a file upload, so the 27 svg
files were uploaded to the database.

A new FoodFetcher class was created to unpack these icon
objects, which contain the filename and a url for the hosted location

The code could be refactored to use the new urls instead
of the fileNames that point to a resource on the webserver
but that is out of scope for this PR.

Furthermore a TreeFetcher class was created so that the fk
food id could be converted from a json encoded array
to a simple string.