Calculate which foods are in season, requires DB update
The functionality was removed when switching to sqlite
It was refactored to be simpler and faster by creating
a new database table to track in_season food.

Originally:
on page load the browser calls calcSeason.php
this would query the donate table to get all foods
that were donated about this time last year or the
year before. From this list of in-season foods it
would then update the entire tree table, ~4000 rows.
For each tree, if its food matched with "in-season"
then a 1 was written to its tree.season column,
otherwise a 0 was written.
The browser would then recieve a list of all trees
and when needed call tree.season to determine if
was needed. Although this was only done in a parent
adoption component.

Now:
An in-season table records whenever a fruit was donated
the day and week of the year it happened.
A simple query can now tell you which foods were donated
during a given week.
The table was seeded with data from the donate table
TODO: it needs to be kept up to date with new donations
entered from the web-app or from airtable directly

seed-data
```sql
INSERT OR IGNORE INTO in_season
SELECT
  SUBSTRING(food, 3, 17) as food_id,
  CAST(STRFTIME('%j', date) as INTEGER) as day_of_year,
  STRFTIME('%W', date) as week_of_year,
  CAST(STRFTIME('%Y', date) as INTEGER) as year
FROM Donate;
```