Add Airtable webhook for Tree table, requires new db and .env
Adds a handler for airtable webhooks
When a change is detected matching the webhooks criteria
Airtable sends a ping to the server with the webhook id
Airtable autosaves often causing many notifications.
For each ping the server requests a list of changes from airtable.
To keep track of what changes the server has already seen a new db,
food-map.sqlite, was created to track the cursor value marking latest
changes the server has seen.
From the response, the server gets a list of unique table and
record ids that were modified. The notification includes the
fieldId and changed value, but the server has no mapping for field
id to column names.
So the server then fetches the the whole modified record.
This is rate limited by airtable at 5 reqs per second.
The newly fetched records are then added to sqlite or deleted if
it was a delete.

The FE browswer is not notified of any changes, but if the user
refreshes the page the browser will see new data