Deploy to tree-map
A number of changes required here as the file path to assets
is hard coded in many files: index.html, scripts, etc.

Also this commit fixes a bug with the admin_check where
admin users could not see private trees even after login.
This was because the admin_check relied on undefined behavior
calling row_count after a SELECT stmt

From PHP docs:
For statements that produce result sets, such as SELECT, the behavior is undefined and can be different for each driver. Some databases may return the number of rows produced by that statement (e.g. MySQL in buffered mode), but this behaviour is not guaranteed for all databases and should not be relied on for portable applications.